format code

+++ src/interfaces/SafeMoon/ISafeswapCallee.sol
@@ -2,5 +2,5 @@
 pragma solidity 0.8.11;
 
 interface ISafeswapCallee {
-    function safeswapCall(address sender, uint amount0, uint amount1, bytes calldata data) external;
+    function safeswapCall(address sender, uint256 amount0, uint256 amount1, bytes calldata data) external;
 }
